Moved buffer tasks into BufferManager Moved parseRuleString into General
